+++ packages/pages/src/common/src/feature/stream.ts
@@ -1,5 +1,12 @@
+import fs from "fs";
+import path from "path";
+import YAML from "yaml";
+
+import { readJsonSync } from "../../../upgrade/migrateConfig.js";
 import { TemplateConfigInternal } from "../template/internal/types.js";
 import { RedirectConfigInternal } from "../redirect/internal/types.js";
+import { ProjectStructure } from "../project/structure.js";
+import { logErrorAndExit } from "../../../util/logError.js";
 
 /**
  * The shape of data that represents a stream configuration.
@@ -29,6 +36,39 @@ export interface StreamConfig {
     /** Whether to include the primary profile language. */
     primary: boolean;
   };
+  /** The transformation to apply to the stream */
+  transform?: {
+    /** The option fields to be expanded to include the display fields, numeric values, and selected boolean */
+    expandOptionFields?: string[];
+    /** The option fields to be replaced with display names */
+    replaceOptionValuesWithDisplayNames?: string[];
+  };
+}
+
+/**
+ * The shape of data that represents a site stream.
+ * Similar to {@link StreamConfig} but the id is always site-stream
+ * and there can only be one entityId.
+ */
+export interface SiteStream {
+  /** Identifies the stream as a site stream */
+  id: "site-stream";
+  /** The entity id of the site stream */
+  entityId: string;
+  /** The fields to apply to the stream */
+  fields: string[];
+  /** The localization used by the filter. Either set primary: true or specify a locales array. */
+  localization:
+    | {
+        /** The entity profiles languages to apply to the stream. */
+        locales: string[];
+        primary?: never;
+      }
+    | {
+        /** Use the primary profile language. */
+        primary: true;
+        locales?: never;
+      };
   /** The transformation to apply to the stream */
   transform?: {
     /** The option fields to be expanded to include the display fields, numeric values, and selected boolean */
@@ -74,4 +114,55 @@ export const convertRedirectConfigToStreamConfig = (
       destination: "pages",
     };
   }
+};
+
+/**
+ * Loads the site stream specified in config.yaml or site-stream.json into a {@link SiteStream}.
+ */
+export const readSiteStream = (
+  projectStructure: ProjectStructure
+): SiteStream | undefined => {
+  // read site stream from deprecated sites-config directory if it exists
+  const siteStreamJsonPath = path.resolve(
+    projectStructure.getSitesConfigPath().path,
+    projectStructure.config.sitesConfigFiles.siteStream
+  );
+  if (fs.existsSync(siteStreamJsonPath)) {
+    const sitesJson = readJsonSync(siteStreamJsonPath);
+    return formatSiteStream(sitesJson, siteStreamJsonPath);
+  }
+
+  // read site stream from config.yaml
+  const configYamlPath = projectStructure.getConfigYamlPath().getAbsolutePath();
+  if (fs.existsSync(configYamlPath)) {
+    const yamlDoc = YAML.parse(fs.readFileSync(configYamlPath, "utf-8"));
+    if (yamlDoc.siteStream) {
+      yamlDoc.siteStream.entityId = yamlDoc.siteStream?.entityId?.toString();
+      return yamlDoc.siteStream;
+    }
+  }
+
+  return;
+};
+
+/**
+ * Converts the deprecated format of a siteStream specified in site-stream.json into
+ * the format of a siteStream specified in config.yaml
+ */
+export const formatSiteStream = (sitesJson: any, siteStreamPath: string) => {
+  let entityId;
+  if (sitesJson.filter?.entityIds && sitesJson.filter?.entityIds.length === 1) {
+    entityId = sitesJson.filter.entityIds[0];
+  } else if (sitesJson.filter?.entityIds) {
+    logErrorAndExit(
+      `Unable to migrate ${siteStreamPath} due to multiple entityIds`
+    );
+  }
+
+  return {
+    id: sitesJson.$id, // Replace $id with id and keeps id in the first position
+    entityId: entityId?.toString(),
+    localization: sitesJson.localization,
+    fields: sitesJson.fields,
+  };
 };
