+++ packages/pages/src/common/src/template/hydration.ts
@@ -1,7 +1,8 @@
 import { HeadConfig, renderHeadConfigToString } from "./head.js";
 import { convertToPosixPath } from "./paths.js";
-import { SiteStreamConfig, TemplateRenderProps } from "./types.js";
+import { TemplateRenderProps } from "./types.js";
 import { FeaturesConfig } from "../feature/features.js";
+import { SiteStream } from "../feature/stream.js";
 
 /**
  * Imports the custom hydration template and entrypoint template as modules and calls
@@ -144,7 +145,7 @@ export const getIndexTemplateDev = (
   indexHtml: string,
   appLanguage: string,
   templatesConfig: FeaturesConfig,
-  siteStream: SiteStreamConfig | undefined,
+  siteStream: SiteStream | undefined,
   headConfig?: HeadConfig
 ): string => {
   let commonIndex = getCommonInjectedIndexHtml(
