+++ packages/pages/src/upgrade/migrateConfig.test.ts
@@ -1,47 +1,5 @@
-import { describe, it, expect, vi } from "vitest";
-import { formatServing, formatSiteStream } from "./migrateConfig.js";
-
-const siteStreamPath = "foo/bar";
-
-describe("formatSiteStream", () => {
-  it("errors and exits when there are multiple entityIds", () => {
-    const testJson = { filter: { entityIds: ["1234", "123"] } };
-    const mockExit = vi
-      .spyOn(process, "exit")
-      .mockImplementation(() => undefined as never);
-    formatSiteStream(testJson, siteStreamPath);
-    expect(mockExit).toHaveBeenCalledWith(1);
-  });
-
-  it("returns expected entityId", () => {
-    const testJson = { filter: { entityIds: ["1234"] } };
-    const expectedJson = { entityId: "1234" };
-    expect(formatSiteStream(testJson, siteStreamPath)).toEqual(expectedJson);
-  });
-
-  it("returns expected id with id first", () => {
-    const testJson = { fields: ["meta", "name"], $id: "123" };
-    const expectedJson = { id: "123", fields: ["meta", "name"] };
-    expect(formatSiteStream(testJson, siteStreamPath)).toEqual(expectedJson);
-  });
-
-  it("returns expected full config", () => {
-    const testJson = {
-      $id: "123",
-      fields: ["meta", "name"],
-      filter: { entityIds: ["1234"] },
-      source: "foo",
-      localization: ["en"],
-    };
-    const expectedJson = {
-      id: "123",
-      entityId: "1234",
-      fields: ["meta", "name"],
-      localization: ["en"],
-    };
-    expect(formatSiteStream(testJson, siteStreamPath)).toEqual(expectedJson);
-  });
-});
+import { describe, it, expect } from "vitest";
+import { formatServing } from "./migrateConfig.js";
 
 describe("formatServing", () => {
   it("returns expected reverseProxyPrefix", () => {
